Initial commit of Fo_effector_clustering pipeline
This is a working version of the pipeline that
has now been committed for the first time. Some
things that I'd like to improve are location of
output directories and making sure the pipeline
file ('Fo_effector_clustering.py') gets copied
to the output folder so that all settings are
saved.